requirements.txt
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.26.148
botocore==1.29.148
certifi==2023.5.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.1
django-allauth==0.54.0
django-crispy-forms==2.0
django-storages==1.13.2
django-widget-tweaks==1.4.12
Faker==18.9.0
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.5
oauthlib==3.2.2
Pillow==9.5.0
platformdirs==3.9.1
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
razorpay==1.3.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.1
virtualenv-clone==0.5.7
websockets==11.0.3
zope.interface==6.0
